Update Swift with Immutable Document API
* Applied Immutable Document API to Swift.
* Refactored Swift unit tests.
