+++ Objective-C/Internal/CBLArray+Swift.h
