+++ Objective-C/Internal/CBLDictionary+Swift.h
