+++ Swift/ArrayObject.swift
@@ -9,548 +9,220 @@
 import Foundation
 
 
-/** ArrayProtocol protocol defines a set of methods for getting and setting array data. */
-protocol ArrayProtocol: ReadOnlyArrayProtocol, ArrayFragment {
+/// ArrayProtocol defines a set of methods for readonly accessing array data.
+protocol ArrayProtocol: ArrayFragment, Sequence {
+    var count: Int { get }
     
-    // MARK: Type Setters
+    func value(at index: Int) -> Any?
     
-    @discardableResult func setArray(_ value: ArrayObject?, at index: Int) -> Self
+    func string(at index: Int) -> String?
     
-    @discardableResult func setBlob(_ value: Blob?, at index: Int) -> Self
+    func int(at index: Int) -> Int
     
-    @discardableResult func setBoolean(_ value: Bool, at index: Int) -> Self
+    func int64(at index: Int) -> Int64
     
-    @discardableResult func setDate(_ value: Date?, at index: Int) -> Self
+    func float(at index: Int) -> Float
     
-    @discardableResult func setDictionary(_ value: DictionaryObject?, at index: Int) -> Self
+    func double(at index: Int) -> Double
     
-    @discardableResult func setDouble(_ value: Double, at index: Int) -> Self
+    func boolean(at index: Int) -> Bool
     
-    @discardableResult func setFloat(_ value: Float, at index: Int) -> Self
+    func blob(at index: Int) -> Blob?
     
-    @discardableResult func setInt(_ value: Int, at index: Int) -> Self
+    func date(at index: Int) -> Date?
     
-    @discardableResult func setInt64(_ value: Int64, at index: Int) -> Self
+    func array(at index: Int) -> ArrayObject?
     
-    @discardableResult func setValue(_ value: Any?, at index: Int) -> Self
+    func dictionary(at index: Int) -> DictionaryObject?
     
-    @discardableResult func setString(_ value: String?, at index: Int) -> Self
-    
-    // MARK: Type Appenders
-    
-    @discardableResult func addArray(_ value: ArrayObject?) -> Self
-    
-    @discardableResult func addBlob(_ value: Blob?) -> Self
-    
-    @discardableResult func addBoolean(_ value: Bool) -> Self
-    
-    @discardableResult func addDate(_ value: Date?) -> Self
-    
-    @discardableResult func addDictionary(_ value: DictionaryObject?) -> Self
-    
-    @discardableResult func addDouble(_ value: Double) -> Self
-    
-    @discardableResult func addFloat(_ value: Float) -> Self
-    
-    @discardableResult func addInt(_ value: Int) -> Self
-    
-    @discardableResult func addInt64(_ value: Int64) -> Self
-    
-    @discardableResult func addValue(_ value: Any?) -> Self
-    
-    @discardableResult func addString(_ value: String?) -> Self
-    
-    // MARK: Type Inserters
-    
-    @discardableResult func insertArray(_ value: ArrayObject?, at index: Int) -> Self
-    
-    @discardableResult func insertBlob(_ value: Blob?, at index: Int) -> Self
-    
-    @discardableResult func insertBoolean(_ value: Bool, at index: Int) -> Self
-    
-    @discardableResult func insertDate(_ value: Date?, at index: Int) -> Self
-    
-    @discardableResult func insertDictionary(_ value: DictionaryObject?, at index: Int) -> Self
-    
-    @discardableResult func insertDouble(_ value: Double, at index: Int) -> Self
-    
-    @discardableResult func insertFloat(_ value: Float, at index: Int) -> Self
-    
-    @discardableResult func insertInt(_ value: Int, at index: Int) -> Self
-    
-    @discardableResult func insertInt64(_ value: Int64, at index: Int) -> Self
-    
-    @discardableResult func insertValue(_ value: Any?, at index: Int) -> Self
-    
-    @discardableResult func insertString(_ value: String?, at index: Int) -> Self
-    
-    // MARK: Removing Value
-    
-    @discardableResult func remove(at index: Int) -> Self
-    
-    // MARK: Setting content with an NSArray
-    
-    @discardableResult func setArray(_ array: Array<Any>?) -> Self
-    
-    // MARK: Getting ArrayObject and DictionaryObject
-    
-    func array(at index: Int) -> ArrayObject? /* override */
-    
-    func dictionary(at index: Int) -> DictionaryObject? /* override */
+    func toArray() -> Array<Any>
 }
 
-/** ArrayObject provides access to array data. */
-public class ArrayObject: ReadOnlyArrayObject, ArrayProtocol {
-    
-    // MARK: Initializers
-    
-    
-    /// Initialize a new empty CBLMutableArray object.
-    public init() {
-        super.init(CBLMutableArray())
-    }
-    
-    
-    /// Initialize a new ArrayObject object with an array content. Allowed value types are NSArray,
-    /// NSDate, NSDictionary, NSNumber, NSNull, NSString, ArrayObject, CBLBlob, DictionaryObject.
-    /// The NSArrays and NSDictionaries must contain only the above types.
-    ///
-    /// - Parameter array: The array object.
-    public init(array: Array<Any>?) {
-        super.init(CBLMutableArray())
-        setArray(array)
-    }
-    
-    
-    // MARK: Type Setters
-    
-    
-    /// Sets an ArrayObject object at the given index. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameters:
-    ///   - value: The ArrayObject object.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func setArray(_ value: ArrayObject?, at index: Int) -> Self {
-        return setValue(value, at: index)
-    }
-    
-    
-    /// Sets a Blob object at the given index. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameters:
-    ///   - value: The Blob object.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func setBlob(_ value: Blob?, at index: Int) -> Self {
-        return setValue(value, at: index)
-    }
-    
-    
-    /// Sets a boolean value at the given index.
-    ///
-    /// - Parameters:
-    ///   - value: The boolean value.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func setBoolean(_ value: Bool, at index: Int) -> Self {
-        return setValue(value, at: index)
-    }
-    
-    
-    /// Sets a Date object at the given index. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameters:
-    ///   - value: The Date object.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func setDate(_ value: Date?, at index: Int) -> Self {
-        return setValue(value, at: index)
-    }
-    
-    
-    /// Sets a DictionaryObject object at the given index. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameters:
-    ///   - value: The DictionaryObject object.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func setDictionary(_ value: DictionaryObject?, at index: Int) -> Self {
-        return setValue(value, at: index)
-    }
-    
-    
-    /// Sets a double value at the given index.
-    ///
-    /// - Parameters:
-    ///   - value: The double value.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func setDouble(_ value: Double, at index: Int) -> Self {
-        return setValue(value, at: index)
-    }
-    
-    
-    /// Sets a float value at the given index.
-    ///
-    /// - Parameters:
-    ///   - value: The float value.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func setFloat(_ value: Float, at index: Int) -> Self {
-        return setValue(value, at: index)
-    }
-    
-    
-    /// Sets an int value at the given index.
-    ///
-    /// - Parameters:
-    ///   - value: The int value.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func setInt(_ value: Int, at index: Int) -> Self {
-        return setValue(value, at: index)
-    }
-    
-    
-    /// Sets an int64 value at the given index.
-    ///
-    /// - Parameters:
-    ///   - value: The int64 value.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func setInt64(_ value: Int64, at index: Int) -> Self {
-        return setValue(value, at: index)
-    }
-    
-    
-    /// Sets a value at the given index. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameters:
-    ///   - value: The value.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func setValue(_ value: Any?, at index: Int) -> Self {
-        arrayImpl.setObject(DataConverter.convertSETValue(value), at: UInt(index))
-        return self
-    }
-    
-    
-    /// Sets a String object at the given index. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameters:
-    ///   - value: The String object.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func setString(_ value: String?, at index: Int) -> Self {
-        return setValue(value, at: index)
-    }
-    
-    
-    // MARK: Type Appenders
-    
-    
-    /// Adds an ArrayObject object to the end of the array. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameter value: The ArrayObject object.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func addArray(_ value: ArrayObject?) -> Self {
-        return addValue(value)
-    }
-    
-    
-    /// Adds a Blob object to the end of the array. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameter value: The Blob object.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func addBlob(_ value: Blob?) -> Self {
-        return addValue(value)
-    }
-    
-    
-    /// Adds a boolean value to the end of the array.
-    ///
-    /// - Parameter value: The boolean value.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func addBoolean(_ value: Bool) -> Self {
-        return addValue(value)
-    }
-    
-    
-    /// Adds a Date object to the end of the array. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameter value: The Date object.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func addDate(_ value: Date?) -> Self {
-        return addValue(value)
-    }
-    
-    
-    /// Adds a Dictionary object to the end of the array. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameter value: The Dictionary object.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func addDictionary(_ value: DictionaryObject?) -> Self {
-        return addValue(value)
-    }
-    
-    
-    /// Adds a double value to the end of the array.
-    ///
-    /// - Parameter value: The double value.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func addDouble(_ value: Double) -> Self {
-        return addValue(value)
-    }
-    
-    
-    /// Adds a float value to the end of the array.
-    ///
-    /// - Parameter value: The double value.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func addFloat(_ value: Float) -> Self {
-        return addValue(value)
-    }
-    
-    
-    /// Adds an int value to the end of the array.
-    ///
-    /// - Parameter value: The integer value.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func addInt(_ value: Int) -> Self {
-        return addValue(value)
-    }
-    
-    
-    /// Adds an int64 value to the end of the array.
-    ///
-    /// - Parameter value: The integer value.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func addInt64(_ value: Int64) -> Self {
-        return addValue(value)
-    }
-    
-    
-    /// Adds a value to the end of the array. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameter value: The value.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func addValue(_ value: Any?) -> Self {
-        arrayImpl.add(DataConverter.convertSETValue(value))
-        return self
-    }
-    
-    
-    /// Adds a String object to the end of the array. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameter value: The String object.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func addString(_ value: String?) -> Self {
-        return addValue(value)
-    }
-    
-    
-    // MARK: Type Inserters
-    
-    
-    /// Inserts an ArrayObject at the given index. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameters:
-    ///   - value: The ArrayObject object.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func insertArray(_ value: ArrayObject?, at index: Int) -> Self {
-        return insertValue(value, at: index)
-    }
-    
-    
-    /// Inserts a Blob object at the given index. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameters:
-    ///   - value: The Blob object.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func insertBlob(_ value: Blob?, at index: Int) -> Self {
-        return insertValue(value, at: index)
-    }
-    
-    
-    /// Inserts a boolean value at the given index.
-    ///
-    /// - Parameters:
-    ///   - value: The boolean value.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func insertBoolean(_ value: Bool, at index: Int) -> Self {
-        return insertValue(value, at: index)
-    }
-    
-    
-    /// Inserts a Date object at the given index. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameters:
-    ///   - value: The Date object.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func insertDate(_ value: Date?, at index: Int) -> Self {
-        return insertValue(value, at: index)
-    }
-    
-    
-    /// Inserts a Dictionary object at the given index. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameters:
-    ///   - value: The Dictionary object.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func insertDictionary(_ value: DictionaryObject?, at index: Int) -> Self {
-        return insertValue(value, at: index)
-    }
-    
-    
-    /// Inserts a double value at the given index. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameters:
-    ///   - value: The double value.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func insertDouble(_ value: Double, at index: Int) -> Self {
-        return insertValue(value, at: index)
-    }
-    
-    
-    /// Inserts a float value at the given index. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameters:
-    ///   - value: The float value.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func insertFloat(_ value: Float, at index: Int) -> Self {
-        return insertValue(value, at: index)
-    }
-    
-    
-    /// Inserts an int value at the given index. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameters:
-    ///   - value: The int value.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func insertInt(_ value: Int, at index: Int) -> Self {
-        return insertValue(value, at: index)
-    }
-    
-    
-    /// Inserts an int64 value at the given index. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameters:
-    ///   - value: The int64 value.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func insertInt64(_ value: Int64, at index: Int) -> Self {
-        return insertValue(value, at: index)
-    }
-    
-    
-    /// Inserts a value at the given index. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameters:
-    ///   - value: The value.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func insertValue(_ value: Any?, at index: Int) -> Self {
-        arrayImpl.insert(DataConverter.convertSETValue(value), at: UInt(index))
-        return self
-    }
-    
-    
-    /// Inserts a String object at the given index. A nil value will be converted to an NSNull.
-    ///
-    /// - Parameters:
-    ///   - value: The String object.
-    ///   - index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func insertString(_ value: String?, at index: Int) -> Self {
-        return insertValue(value, at: index)
-    }
-    
-    
-    // MARK: Removing Value
-    
-    
-    /// Removes the object at the given index.
-    ///
-    /// - Parameter index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func remove(at index: Int) -> Self {
-        arrayImpl.removeObject(at: UInt(index))
-        return self
-    }
-    
-    
-    // MARK: Setting content with an NSArray
-    
 
-    /// Set an array as a content. Allowed value types are Array, Date, Dictionary, Number,
-    /// NSNull, String, ArrayObject, Blob, DictionaryObject. The Array and Dictionary must
-    /// contain only the above types. Setting the new array content will replace the current data
-    /// including the existing ArrayObject and DictionaryObject objects.
-    ///
-    /// - Parameter array: The array.
-    /// - Returns: The ArrayObject object.
-    @discardableResult public func setArray(_ array: Array<Any>?) -> Self {
-        arrayImpl.setArray(DataConverter.convertSETArray(array))
-        return self
+/// ArrayObject provides readonly access to array data.
+public class ArrayObject: ArrayProtocol {
+    
+    /// Gets a number of the items in the array.
+    public var count: Int {
+        return Int(_impl.count)
     }
     
     
-    // MARK: Getting ArrayObject and DictionaryObject
-    
-    
-    /// Gets an ArrayObject at the given index. Returns nil if the value is not an array.
+    /// Gets the value at the given index as a ArrayObject value, which is a mapping object
+    /// of an array value.
+    /// Returns nil if the value doesn't exists, or its value is not an array.
     ///
-    /// - Parameter index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The ArrayObject object.
-    public override func array(at index: Int) -> ArrayObject? {
+    /// - Parameter index: The index.
+    /// - Returns: The ArrayObject at the given index.
+    public func array(at index: Int) -> ArrayObject? {
         return value(at: index) as? ArrayObject
     }
     
     
-    /// Gets a DictionaryObject at the given index. Returns nil if the value is not a dictionary.
+    /// Get the Blob value at the given index.
+    /// Returns nil if the value doesn't exist, or its value is not a blob.
     ///
-    /// - Parameter index: The index. This value must not exceed the bounds of the array.
-    /// - Returns: The DictionaryObject object.
-    public override func dictionary(at index: Int) -> DictionaryObject? {
-        return value(at: index) as? DictionaryObject
+    /// - Parameter index: The index.
+    /// - Returns: The Blob value located at the index.
+    public func blob(at index: Int) -> Blob? {
+        return _impl.blob(at: UInt(index))
+    }
+
+    
+    /// Gets the value at the given index as a boolean value.
+    /// Returns true if the value exists, and is either `true` or a nonzero number.
+    ///
+    /// - Parameter index: The index.
+    /// - Returns: The Bool value located at the index.
+    public func boolean(at index: Int) -> Bool {
+        return _impl.boolean(at: UInt(index))
     }
     
     
-    // MARK: ArrayFragment
-    
-    
-    /// Subscripting access to a Fragment object that represents the value at the given index.
+    /// Gets value at the given index as an Date.
+    /// JSON does not directly support dates, so the actual property value must be a string, which
+    /// is then parsed according to the ISO-8601 date format (the default used in JSON.)
+    /// Returns nil if the value doesn't exist, is not a string, or is not parseable as a date.
+    /// NOTE: This is not a generic date parser! It only recognizes the ISO-8601 format, with or
+    /// without milliseconds.
     ///
-    /// - Parameter index: The index. If the index value exceeds the bounds of the array,
-    ///                    the Fragment object will represent a nil value.
-    /// - Returns: The Fragment object.
-    public override subscript(index: Int) -> Fragment {
-        return Fragment(arrayImpl[UInt(index)])
+    /// - Parameter index: The index.
+    /// - Returns: The Date value at the given index.
+    public func date(at index: Int) -> Date? {
+        return _impl.date(at: UInt(index))
+    }
+    
+    
+    /// Get the value at the given index as a DictionaryObject value, which is a
+    /// mapping object of a dictionary value.
+    /// Returns nil if the value doesn't exists, or its value is not a dictionary.
+    ///
+    /// - Parameter index: The index.
+    /// - Returns: The DictionaryObject
+    public func dictionary(at index: Int) -> DictionaryObject? {
+        return value(at: index) as? DictionaryObject
+    }
+
+    
+    /// Gets the value at the given index as a float value.
+    /// Integers will be converted to float. The value `true` is returned as 1.0, `false` as 0.0.
+    //// Returns 0.0 if the value doesn't exist or does not have a numeric value.
+    ///
+    /// - Parameter index: The index.
+    /// - Returns: The Float value located at the index.
+    public func float(at index: Int) -> Float {
+        return _impl.float(at: UInt(index))
+    }
+    
+    
+    /// Gets the value at the given index as a double value.
+    /// Integers will be converted to double. The value `true` is returned as 1.0, `false` as 0.0.
+    /// Returns 0.0 if the property doesn't exist or does not have a numeric value.
+    ///
+    /// - Parameter index: The index.
+    /// - Returns: The Double value located at the index.
+    public func double(at index: Int) -> Double {
+        return _impl.double(at: UInt(index))
+    }
+    
+    
+    /// Gets value at the given index as an integer value.
+    /// Floating point values will be rounded. The value `true` is returned as 1, `false` as 0.
+    /// Returns 0 if the value doesn't exist or does not have a numeric value.
+    ///
+    /// - Parameter index: The index.
+    /// - Returns: The integer value located at the index.
+    public func int(at index: Int) -> Int {
+        return _impl.integer(at: UInt(index))
+    }
+    
+    
+    /// Gets value at the given index as an integer value.
+    /// Floating point values will be rounded. The value `true` is returned as 1, `false` as 0.
+    /// Returns 0 if the value doesn't exist or does not have a numeric value.
+    ///
+    /// - Parameter index: The index.
+    /// - Returns: The integer value located at the index.
+    public func int64(at index: Int) -> Int64 {
+        return _impl.longLong(at: UInt(index))
+    }
+
+
+    /// Gets the value at the given index as a string.
+    /// Returns nil if the value doesn't exist, or its value is not a string
+    ///
+    /// - Parameter index: The index.
+    /// - Returns: The String object.
+    public func string(at index: Int) -> String? {
+        return _impl.string(at: UInt(index))
+    }
+    
+    
+    /// Gets the value at the given index. The value types are Blob, ArrayObject,
+    /// DictionaryObject, Number, or String based on the underlying data type.
+    ///
+    /// - Parameter index: The index.
+    /// - Returns: The value located at the index.
+    public func value(at index: Int) -> Any? {
+        return DataConverter.convertGETValue(_impl.object(at: UInt(index)))
+    }
+    
+    
+    /// Gets content of the current object as an Array object. The values contained in the
+    /// returned Array object are all JSON based values.
+    ///
+    /// - Returns: The Array object representing the content of the current object in
+    ///            the JSON format.
+    public func toArray() -> Array<Any> {
+        return _impl.toArray()
+    }
+    
+    
+    // MARK: Sequence
+    
+    
+    /// Gets an iterator over items in the array.
+    ///
+    /// - Returns: The array item iterator.
+    public func makeIterator() -> AnyIterator<Any> {
+        var index = 0;
+        let count = self.count
+        return AnyIterator {
+            if index < count {
+                let v = self.value(at: index)
+                index += 1
+                return v;
+            }
+            return nil
+        }
+    }
+    
+    
+    // MARK: Subscript
+    
+    
+    /// Subscript access to a Fragment object by index
+    ///
+    /// - Parameter index: The Index.
+    public subscript(index: Int) -> Fragment {
+        return Fragment(_impl[UInt(index)])
     }
     
     
     // MARK: Internal
     
     
-    init(_ impl: CBLMutableArray) {
-        super.init(impl)
+    init(_ impl: CBLArray) {
+        _impl = impl
+        _impl.swiftObject = self
     }
     
     
-    // MARK: Private
-    
-    
-    private var arrayImpl: CBLMutableArray {
-        return _impl as! CBLMutableArray
+    deinit {
+        _impl.swiftObject = nil
     }
+    
+    
+    let _impl: CBLArray
     
 }
