+++ Swift/CouchbaseLiteSwift.h
@@ -15,6 +15,6 @@ FOUNDATION_EXPORT double CouchbaseLiteSwiftVersionNumber;
 FOUNDATION_EXPORT const unsigned char CouchbaseLiteSwiftVersionString[];
 
 #import "CouchbaseLite.h"
-#import "CBLPredicateQuery.h"
-#import "CBLReadOnlyArray+Swift.h"
+#import "CBLArray+Swift.h"
 #import "CBLDictionary+Swift.h"
+#import "CBLPredicateQuery.h"
