+++ Swift/DataConverter.swift
@@ -16,22 +16,22 @@ import Foundation
             if let dict = implDict.swiftObject {
                 return dict
             }
-            return DictionaryObject(implDict)
+            return MutableDictionaryObject(implDict)
         case let implArray as CBLMutableArray:
             if let array = implArray.swiftObject {
                 return array
             }
-            return ArrayObject(implArray)
+            return MutableArrayObject(implArray)
         case let implDict as CBLDictionary:
             if let dict = implDict.swiftObject {
                 return dict
             }
-            return ReadOnlyDictionaryObject(implDict)
+            return DictionaryObject(implDict)
         case let implArray as CBLArray:
             if let array = implArray.swiftObject {
                 return array
             }
-            return ReadOnlyArrayObject(implArray)
+            return ArrayObject(implArray)
         default:
             return value
         }
@@ -40,9 +40,9 @@ import Foundation
     
     static func convertSETValue(_ value: Any?) -> Any? {
         switch value {
-        case let dict as DictionaryObject:
+        case let dict as MutableDictionaryObject:
             return dict._impl
-        case let array as ArrayObject:
+        case let array as MutableArrayObject:
             return array._impl
         default:
             return value
